Created conversationInstructions UI on Chat
The main section of the chat component will show a message to the user based on whether the user is part of any conversations.

If there are conversations present, the message will say " to get started, select a conversation".
If there are no conversations present, The message will state "to get started, create a conversation from the sidebar"